Split connection and client classes
This is a breaking change, API-wise. The library no longer exposes the simple
`EM::APN.push` method. Instead, an `EM::APN::Client` object is instantiated
and `EM::APN::Notification` objects are passed to #deliver.

Error callbacks can be attached to the Client instance, as well as callbacks
when the connection closes.